CheckResizer_Complete_Offline/requirements_offline.txt
streamlit
opencv-python  
pillow
numpy
scipy
scikit-learn
matplotlib
